.prepareRasterBackedFile -- disallow overwrites if coming from new save
This is the fix for a long standing bug where attempts to recover raster from cache fail because the file backing the raster is gone. This was diagnosed: it is when 2 functions output same file, then one is deleted via clearCache. The other cache event is still there, but its file-backed component is gone.

+++ R/cache-helpers.R
@@ -259,7 +259,9 @@ setMethod(
   ".prepareOutput",
   signature = "RasterLayer",
   definition = function(object, cacheRepo, ...) {
-    .prepareFileBackedRaster(object, repoDir = cacheRepo)
+    # with this call to .prepareFileBackedRaster, it is from the same function call as a previous time
+    #  overwrite is ok
+    .prepareFileBackedRaster(object, repoDir = cacheRepo, overwrite = TRUE, ...)
 })
 
 #' @export
@@ -562,6 +564,8 @@ setMethod(
 #'
 #' @param repoDir Character denoting an existing directory in which an artifact will be saved.
 #'
+#' @param overwrite Logical. Should the raster be saved to disk, overwriting existing file.
+#'
 #' @param ... passed to \code{archivist::saveToRepo}
 #'
 #' @return A raster object and its newly located file backing.
@@ -590,7 +594,7 @@ setMethod(
 #' r # now in "rasters" subfolder of tempdir()
 #'
 #'
-.prepareFileBackedRaster <- function(obj, repoDir = NULL) {
+.prepareFileBackedRaster <- function(obj, repoDir = NULL, overwrite = FALSE, ...) {
   isRasterLayer <- TRUE
   isStack <- is(obj, "RasterStack")
   repoDir <- checkPath(repoDir, create = TRUE)
@@ -629,12 +633,13 @@ setMethod(
       normalizePath(file.path(repoDir, splittedFilenames2), winslash = "/", mustWork = FALSE)
     }
     if (any(!file.exists(trySaveFilename))) {
-      stop("The following rasters are supposed to be on disk but appear to have been deleted:\n",
+      stop("The following file-backed rasters are supposed to be on disk ",
+           "but appear to have been deleted:\n",
            paste("    ", curFilename, collapse = "\n"),
-           "\n\nThese files must be recreated, e.g., try:\n",
-           "  showCache(userTags = \"writeRaster\").\n",
-           "\nExamine that and possibly [with caution!]:\n",
-           "  reproducible::clearCache(userTags = \"writeRaster\")")
+           "The most likely reason is that two functions had the same output ",
+           "and one of them was removed with clearCache(...). ",
+           "The best solution to this is never have two functions create the same ",
+           "file-backed raster.")
     } else {
       slot(slot(obj, "file"), "name") <- saveFilename <- curFilename <- trySaveFilename
     }
@@ -648,6 +653,20 @@ setMethod(
     saveFilename <- normalizePath(saveFilename, winslash = "/", mustWork = FALSE)
   }
 
+  sameFilenames <- saveFilename == curFilename
+  if (any(sameFilenames)) {
+    if (!overwrite) {
+      saveFilename[sameFilenames] <- unlist(lapply(seq_along(curFilename[sameFilenames]),
+             function(x) {
+               if (file.exists(saveFilename[x])) {
+                 suff <- paste0("_", paste(collapse="", sample(LETTERS, 5)))
+                 .suffix(saveFilename[x], suff)
+               } else {
+                 saveFilename[x]
+               }
+             }))
+    }
+  }
   # filenames are not the same
   if (any(saveFilename != curFilename)) {
     if (isFilebacked) {
@@ -659,17 +678,24 @@ setMethod(
             unique() %>%
             sapply(., dir.create, recursive = TRUE)
         }
+
         if (any(saveFilename %>% grepl(., pattern = "[.]grd$"))) {
           copyFile(from = curFilename, to = saveFilename, overwrite = TRUE, silent = TRUE)
           griFilename <- sub(saveFilename, pattern = "[.]grd$", replacement = ".gri")
           curGriFilename <- sub(curFilename, pattern = "[.]grd$", replacement = ".gri")
           copyFile(from = curGriFilename, to = griFilename, overwrite = TRUE, silent = TRUE)
         } else {
-          suppressWarnings(
+          #suppressWarnings(
             lapply(seq_along(curFilename),
-                   function(x) copyFile(to = saveFilename[x],
+                   function(x) {
+                     if (file.exists(saveFilename[x])) {
+                       suff <- paste0("_", paste(collapse="", sample(LETTERS, 5)))
+                       saveFilename[x] <- .suffix(saveFilename[x], suff)
+                     }
+                     copyFile(to = saveFilename[x],
                                         overwrite = TRUE,
-                                        from = curFilename[x], silent = TRUE)))
+                                        from = curFilename[x], silent = TRUE)
+                   })#)
         }
       }
       # for a stack with independent Raster Layers (each with own file)
@@ -753,49 +779,56 @@ copyFile <- function(from = NULL, to = NULL, useRobocopy = TRUE,
                      #copyRasterFile=TRUE, clearRepo=TRUE,
                      create = TRUE, silent = FALSE) {
   origDir <- getwd()
-  useFileCopy <- FALSE
+  useFileCopy <- identical(dirname(from), dirname(to))
 
   checkPath(dirname(to), create = create)
 
   os <- tolower(Sys.info()[["sysname"]])
-  if (os == "windows") {
-    if (!dir.exists(to)) to <- dirname(to) # extract just the directory part
-    robocopyBin <- tryCatch(Sys.which("robocopy"), warning = function(w) NA_character_)
+  if (!useFileCopy) {
+    if (os == "windows") {
+      if (!dir.exists(to)) toDir <- dirname(to) # extract just the directory part
+      robocopyBin <- tryCatch(Sys.which("robocopy"), warning = function(w) NA_character_)
 
-    robocopy <-  if (silent) {
-      paste0(robocopyBin, " /purge"[delDestination], " /ETA /XJ /XO /NDL /NFL /NJH /NJS \"",  # nolint
-             normalizePath(dirname(from), mustWork = TRUE, winslash = "\\"), "\" \"",
-             normalizePath(to, mustWork = FALSE, winslash = "\\"),  "\" ",
-             basename(from))
-    } else {
-      paste0(robocopyBin, " /purge"[delDestination], " /ETA /XJ /XO \"", # nolint
-             normalizePath(dirname(from), mustWork = TRUE, winslash = "\\"), "\" \"",
-             normalizePath(to, mustWork = FALSE, winslash = "\\"), "\" ",
-             basename(from))
-    }
+      robocopy <-  if (silent) {
+        paste0(robocopyBin, " /purge"[delDestination], " /ETA /XJ /XO /NDL /NFL /NJH /NJS \"",  # nolint
+               normalizePath(dirname(from), mustWork = TRUE, winslash = "\\"), "\" \"",
+               normalizePath(toDir, mustWork = FALSE, winslash = "\\"),  "\" ",
+               basename(from))
+      } else {
+        paste0(robocopyBin, " /purge"[delDestination], " /ETA /XJ /XO \"", # nolint
+               normalizePath(dirname(from), mustWork = TRUE, winslash = "\\"), "\" \"",
+               normalizePath(toDir, mustWork = FALSE, winslash = "\\"), "\" ",
+               basename(from))
+      }
 
-    useFileCopy <- if (useRobocopy && !is.na(robocopyBin)) {
-      suppressWarnings(tryCatch(system(robocopy, intern = TRUE), error = function(x) TRUE))
+      useFileCopy <- if (useRobocopy && !is.na(robocopyBin)) {
+        suppressWarnings(tryCatch(system(robocopy, intern = TRUE), error = function(x) TRUE))
+      } else {
+        TRUE
+      }
+      if (isTRUE(any(grepl("ERROR", useFileCopy)))) {
+        useFileCopy <- TRUE
+      }
+
+      # means that the file didn't get copied because it is actually the same directory
+      if (!nzchar(useFileCopy)) {
+        useFileCopy <- TRUE
+      }
+    } else if ( (os == "linux") || (os == "darwin") ) { # nolint
+      if (!dir.exists(to)) toDir <- dirname(to) # extract just the directory part
+      rsyncBin <- tryCatch(Sys.which("rsync"), warning = function(w) NA_character_)
+      opts <- if (silent) " -a " else " -avP "
+      rsync <- paste0(rsyncBin, " ", opts, " --delete "[delDestination],
+                      normalizePath(from, mustWork = TRUE), " ",
+                      normalizePath(toDir, mustWork = FALSE), "/")
+
+      useFileCopy <- tryCatch(system(rsync, intern = TRUE), error = function(x) TRUE)
     } else {
-      TRUE
-    }
-    if (isTRUE(any(grepl("ERROR", useFileCopy)))) {
       useFileCopy <- TRUE
     }
-  } else if ( (os == "linux") || (os == "darwin") ) { # nolint
-    if (!dir.exists(to)) to <- dirname(to) # extract just the directory part
-    rsyncBin <- tryCatch(Sys.which("rsync"), warning = function(w) NA_character_)
-    opts <- if (silent) " -a " else " -avP "
-    rsync <- paste0(rsyncBin, " ", opts, " --delete "[delDestination],
-                    normalizePath(from, mustWork = TRUE), " ",
-                    normalizePath(to, mustWork = FALSE), "/")
-
-    useFileCopy <- tryCatch(system(rsync, intern = TRUE), error = function(x) TRUE)
-  } else {
-    useFileCopy <- TRUE
   }
   if (isTRUE(useFileCopy)) {
-    dir.create(dirname(to))
+    checkPath(dirname(to), create = TRUE)
     file.copy(from = from, to = to, overwrite = overwrite, recursive = FALSE)
   }
